LayoutTests/crypto/subtle/aes-cbc-import-jwk-key-non-extractable-expected.txt

Test importing a JWK AES-CBC non-extractable key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS key.type is 'secret'
PASS key.extractable is false
PASS key.algorithm.name is 'AES-CBC'
PASS key.algorithm.length is 128
PASS key.usages is ['decrypt', 'encrypt', 'unwrapKey', 'wrapKey']
PASS successfullyParsed is true

TEST COMPLETE
